refactor(#137/*): curation success code global/response 로 이동
